# [onert] Optimize Bias Grad Computation