This commit optimizes bias grad computation with multithreading.

ONE-DCO-1.0-Signed-off-by: YongHyun An <yonghyunz.an@samsung.com>

---
From draft https://github.com/Samsung/ONE/pull/12571, https://github.com/Samsung/ONE/pull/12666
Relevant issue: https://github.com/Samsung/ONE/issues/12569