@jyoungyun 2024-02-22
@Aeren1564 Would you like to look at another file structure? This is because this PR doesn't fit with the existing code design. I think this  `reduxFunctor.h` should be located in the `compute/cker/include/cker/eigen/`. And when copying code from other repository, it is recommended not to modify it as much as possible. So it would be better to keep the file name as before. :)